requirements.txt
asgiref==3.3.4
Django==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
random-name==0.1.1
six==1.16.0
sqlparse==0.4.1
